requirements.txt
django-solo==1.1.2